ref: take enforce limits out of relayer
